\docType{data}
\name{contaminatedTriangles}
\alias{contaminatedTriangles}
\title{
contaminatedTriangles
}
\description{
Simulated triangle data produced as an example of mild contaminated data with
behavioural outliers.
}
\value{
  \item{fd}{A functional data object representing the fitted triangle data.}
  \item{groupd}{Group classifications for each curve as a ordinary behavioral group (1,2,3, or 4) or outlier (4 as outliers to group 1 and 6 as outliers for group 3).}
}
\references{
- Cristina Anton, Iain Smith Model-based clustering of functional data via mixtures of \eqn{t} distributions. Advances in Data Analysis and Classification (to appear).
}
\author{
Cristina Anton and Iain Smith
}
\keyword{datasets}